Create paperScript
This changes the cursor to a different cursor when the mouse is set over it. If you aren't using a custom cursor just remove the custom cursor stuff and use it as a derpy item pick up with the paper Check script. May cause recursor and crash if you set audio listener and source on the same item when you active. If you aren't adding a custom pick up noise don't worry about that.